migrate journeyapps-parser-schema unit tests to vitest
