--- journeyapps-parser-schema/karma.conf.js
@@ -1,1 +1,0 @@
-module.exports = require('../karma.conf')(__dirname);
