--- journeyapps-parser-schema/test/unit/all.ts
@@ -1,3 +1,0 @@
-import '@journeyapps/core-test-helpers';
-import './schemaSpec';
-import './schemaSerializeSpec';
