--- journeyapps-parser-schema/test/webpack.config.js
@@ -1,2 +1,0 @@
-const p = require('path');
-module.exports = require('@journeyapps/core-test-helpers/src/webpack.config')(p.join(__dirname, '..'));
