+++ journeyapps-parser-schema/tests/fixtures/schema2.xml
