+++ journeyapps-parser-schema/tests/fixtures/schema3-badge-count.xml
