+++ journeyapps-parser-schema/tests/fixtures/schema3.xml
