+++ journeyapps-parser-schema/vitest.config.unit.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    include: ['./**/*.test.ts'],
+    reporters: 'verbose'
+  }
+});
