+++ package.json
@@ -4,8 +4,7 @@
   "private": true,
   "workspaces": [
     "journeyapps",
-    "journeyapps-*",
-    "test-helpers"
+    "journeyapps-*"
   ],
   "scripts": {
     "build": "pnpm -r build",
@@ -46,10 +45,5 @@
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1",
     "webpack-node-externals": "^3.0.0"
-  },
-  "pnpm": {
-    "overrides": {
-      "@journeyapps/core-test-helpers": "link:./test-helpers"
-    }
   }
 }
