--- test-helpers/CHANGELOG.md
@@ -1,7 +1,0 @@
-# @journeyapps/core-test-helpers
-
-## 4.1.2
-
-### Patch Changes
-
-- cc89366: Version bump hotfix
