--- test-helpers/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@journeyapps/core-test-helpers",
-  "main": "./src/helpers.js",
-  "version": "4.1.2",
-  "private": true
-}
